+++ CHANGELOG.md
@@ -18,10 +18,13 @@ situations.
 - Added resource limiter for HUMAnN2 due to its intense use of huge temporary
   files in the output folder. Activated with --resources humann2=X, where X is
   the max number of parallel instances of humann2 to run.
+
 ### Changed
 - Added read length window filter before groot alignment step.
 - Change logdir of remove_human rule to LOGDIR/remove_human instead of
   OUTDIR/logs/remove_human.
+- Improved make_count_table.py so it can use TSV annotation files with multiple
+  columns. Added config setting for which columns to include.
 
 
 ## [0.3.2-dev]